Rename and reorg some packages:
Improve understandability and limit
the top level packages to more core
pkgs.
