+++ smee/internal/dhcp/handler/reservation/option.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/insomniacslk/dhcp/dhcpv4"
 	"github.com/tinkerbell/tinkerbell/data"
-	"github.com/tinkerbell/tinkerbell/otel"
+	"github.com/tinkerbell/tinkerbell/pkg/otel"
 	"github.com/tinkerbell/tinkerbell/smee/internal/dhcp"
 	dhcpotel "github.com/tinkerbell/tinkerbell/smee/internal/dhcp/otel"
 )
